feat: add production batch CRUD with models and quantity converter
- Add QuantityConverter helper for bag/pound conversions
- Create Dispatch, DispatchItem, ProductionBatch, ProductionInput, ProductionOutput, and StockBalance models
- Update stock ledger and paddy receipt items quantity columns from decimal to integer
- Make production batch status nullable via migration
- Add production batch API endpoints with filtering and pagination
- Implement store, update, and show operations with bag/loose pound handling

+++ database/migrations/2026_02_20_000003_create_stock_ledger_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->uuid('owner_id');
             $table->uuid('item_id');
             $table->uuid('location_id');
-            $table->decimal('quantity', 12, 2);
+            $table->integer('quantity');
             $table->integer('direction');
             $table->timestamp('created_at')->useCurrent();
 
@@ -36,4 +36,3 @@ return new class extends Migration
         Schema::dropIfExists('stock_ledger');
     }
 };
-
